Moved PythonHelper to PythonSetup as NUnit SetUpFixture
- also made a few minor changes in StandardDeviationExecutionModelTests
